WIP Add images in MDX support
I stopped work on this for 2.5 months and left these changes unstaged :(

If I recall correctly: I was trying to pass MDX with images from
the content app, into the store, and then retrieve from matt.fyi. I
don't believe it was working.

It's possible some parts of this code work and I might even need or
want them.