Add category usecase to graph mapper
Simplifies Category children to be an array of type Category rather
than a union of Category and Risk. This simplifies the mapping for
now, until that union (Concern) is supported in the domain model object.

Disabled schema stitching in graph code generator, which will simplify
typing for resolvers and make unit testing easier.