Add Box component with stories
Adds Box component as a wrapper for Grommet's Box with a subset of
propereties.

Includes stories for the majority of props, with detailed stories for
props that don't have a simple type mapping to Grommet, due to unions
with `string` widening the type to any possible string.

Includes styled-components dependency for stories.

Adds support for dark mode to ThemeProvider.