Separate build from deploy CD jobs and temporary disable deploy
Modifies the Github workflow to build in parallel to running tests and
deploy once the build and test jobs are complete.

Temporarily disables the actual deployment step of the deploy job and
adds a step to print the directory, for testing that the upload and
download works as expected between the build and deploy jobs.