Remove attempt at fetching job information in workflow
The job run instance ID isn't available in context, so it's not possible
to retrieve the job information inside the workflow. Unless, of course,
we are willing to do yet another request to fetch a list of jobs. But
that seems excessively complex.

Adds an attempt to calculate the time as separate steps using
scripts.