Use 50% of screen for light and dark modes in stories
Previously the dark and light columns used only the width they needed.
But when making a change to a small part of a story that increased the
width, this made the whole diff have changes highlighting everywhere
the width of the column changed.

This should give some buffer to increase the width of stories a little
without causing the diff to change significantly