Serialize Date scalar as ISO 8601 string
Changes serializing/deserializing of Date scalar to be an ISO 8601
string intead of a unix epoch number. This is easier to read in graph
queries.

Prints the update dates on the home page to en-US format.

Maps store's Date scalar in Gatsby types to a string, to accurately
represent this change. Not sure if it's possible to deserialize to an
actual Date object.