Target workspaces individually for linting
Changes Typescript and ESLint configs to target workspaces individually
for linting, instead of being run across the entire monorepo. This will
allow Cypress to be added with a separate config and avoid global
conflicts with Jest function names, and also make individual linting
jobs smaller so they don't hit out of memory errors.

Renames util lib to "helper" to remove a warning about there being an
NPM package of that name.

Adds src folder to all workspaces that were missing it, so Typescript
configs can point to that instead of the base directory and avoid errors
on root config files.

Puts an ESLint and Typescript config file in every package for
consistency and granular control.