Add href and raisePosition props to Button component
raisePosition provides a basic way to lift a component above elements
that are layered over it in the DOM. The needed usecase is a Diagram,
where the generated SVG appears after the Buttons in the DOM and thus
cover them, making them not clickable. This will allow them to be
clickable while the diagram will still be visible.