Add loading guard to category link E2E tests
Tells Cypress to wait until the diagram is rendered in the category
navigation on the index page. This prevents a race condition where
Cypress would find the link to be clicked, then Grommet would remove the
links from the DOM to draw the SVG as part of the Diagram component,
which prevented Cypress from clicking the link.

Also removes the `id` prop from the Diagram component as Grommet does
not forward it to the DOM element.